Refactor: split ArenaCameraNode.cpp into focused modules
Reduces the monolithic ArenaCameraNode.cpp from 2334 to 1576 lines (-32%)
by extracting cohesive groups of functions into separate files:

- config_helpers.h: inline YAML config reader helpers (config_string, config_bool, etc.)
- pixel_format_helpers.h: PFNC constants and format detection/naming utilities
- camera_config.cpp: all set_nodes_* camera GenICam configuration methods
- diagnostics.cpp: produce_diagnostics_() ROS diagnostics reporting

Also fixes the QoS translation file:
- Renamed quilty_of_service_translation.cpp → quality_of_service_translation.h
  (fixes spelling typo, converts #include'd .cpp to proper .h)
- Fixed is_supported_qos_histroy_policy → is_supported_qos_history_policy typo

Test files updated to use shared headers instead of duplicated code.
All 24 unit tests pass.
