+++ ros2_ws/src/arena_camera_node/CMakeLists.txt
@@ -117,9 +117,10 @@ set(target_name start)
 add_executable(${target_name}
     src/main.cpp
     src/ArenaCameraNode.cpp
+    src/camera_config.cpp
+    src/diagnostics.cpp
     src/light_arena/deviceinfo_helper.cpp
     src/rclcpp_adapter/pixelformat_translation.cpp
-    src/rclcpp_adapter/quilty_of_service_translation.cpp
 )
 
 target_include_directories(${target_name}
@@ -276,6 +277,9 @@ if(BUILD_TESTING)
   # Config helper and pixel format utility tests
   ament_add_gtest(test_config_helpers
     test/test_config_helpers.cpp
+  )
+  target_include_directories(test_config_helpers PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/src
   )
   target_link_libraries(test_config_helpers yaml-cpp)
 endif()
